neuralagentAI/backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlmodel
pydantic
alembic
sqlalchemy
psycopg2-binary
passlib
bcrypt
pyjwt
boto3
botocore
python-multipart
pillow
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
langchain-aws
langchain-azure-ai
langchain-anthropic
anthropic
youtube-transcript-api
beautifulsoup4
anthropic
transformers
websockets
broadcaster
broadcaster[redis]
asyncio-redis
requests
